Renamed meal_planner app to core
