+++ dev_server/settings.py
@@ -37,7 +37,7 @@ INSTALLED_APPS = [
     'django.contrib.sessions',
     'django.contrib.messages',
     'django.contrib.staticfiles',
-    'meal_planner',
+    'core',
 ]
 
 MIDDLEWARE = [
