+++ dev_server/urls.py
@@ -15,7 +15,7 @@ Including another URLconf
 """
 from django.contrib import admin
 from django.urls import path
-from meal_planner import views
+from core import views
 
 urlpatterns = [
     path('admin/', admin.site.urls),
